+++ cmd/cub3D/movement/update_coodinate.c
@@ -6,27 +6,27 @@
 /*   By: saraki <saraki@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/18 07:43:41 by syamasaw          #+#    #+#             */
-/*   Updated: 2024/10/28 11:05:24 by saraki           ###   ########.fr       */
+/*   Updated: 2024/10/28 16:56:12 by saraki           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-static void	set_player_to_edge(t_dda *dda_ret, t_data *data, int mv_angle);
+static void	set_player_to_edge(t_dda *dda_ret, t_data *data, double mv_angle);
 
 void	update_coordinate(t_data *data, int key)
 {
-	int			angle;
+	double		angle;
 	t_dda		dda_ret;
 	t_player	adjust_player;
 
 	angle = data->player.angle;
 	if (key == KEY_D)
-		angle += 90;
+		angle += 90.0;
 	else if (key == KEY_A)
-		angle -= 90;
+		angle -= 90.0;
 	else if (key == KEY_S)
-		angle += 180;
+		angle += 180.0;
 	angle = regulate_angle(angle);
 	adjust_player.pos_x = data->player.pos_x;
 	adjust_player.pos_y = data->player.pos_y;
@@ -40,7 +40,7 @@ void	update_coordinate(t_data *data, int key)
 	set_player_to_edge(&dda_ret, data, angle);
 }
 
-static void	set_player_to_edge(t_dda *dda, t_data *d, int mv_angle)
+static void	set_player_to_edge(t_dda *dda, t_data *d, double mv_angle)
 {
 	double		next_distance;
 	double		next[2];
@@ -48,9 +48,9 @@ static void	set_player_to_edge(t_dda *dda, t_data *d, int mv_angle)
 	next[0] = STRIDE * cos_degree(mv_angle);
 	next[1] = STRIDE * sin_degree(mv_angle);
 	if (dda->side == 0)
-		next_distance = dda->dist_option[0] - dda->delta_dist[0] - OFFSET; // 省略可能
+		next_distance = dda->dist_option[0] - dda->delta_dist[0] - OFFSET;
 	else
-		next_distance = dda->dist_option[1] - dda->delta_dist[1] - OFFSET; // 省略可能
+		next_distance = dda->dist_option[1] - dda->delta_dist[1] - OFFSET;
 	if (next_distance > sqrt(next[0] * next[0] + next[1] * next[1]))
 	{
 		d->player.pos_x += next[0];
